Add an option to delimit instructions in the trace
Supports both the OCaml and C emulators. For the C emulator this is
enabled with `--trace=step`, and for OCaml `-trace-delimit-step`.
It works slightly differently in both cases, because of how the two
emulators handle the toplevel loop differently.

The C simulator --trace option had it's error message formatted
oddly when the passed option was unrecognised, with the expected
and provided argument flipped, so that has also been fixed.